rename file AngularGaussianPotential to Periodic
