+++ mjolnir/potential/local/PeriodicGaussianPotential.hpp
